improvements to audio splitting -- throws out more that have bad phonemes, fix for race condition on getting back exercise lists, refactoring
Former-commit-id: 52ba9014878f7219ab95486a132a608f5d7b8ba1 [formerly 7c89c7748e752109f2071defe92db19d360df72b] [formerly 4316bb229b7272e4fe0bc2c96f96a341103fecdd [formerly 3455ee0d5944993e22fcec3145632b8594ba581b]] [formerly 52ba9014878f7219ab95486a132a608f5d7b8ba1 [formerly 7c89c7748e752109f2071defe92db19d360df72b] [formerly 4316bb229b7272e4fe0bc2c96f96a341103fecdd [formerly 3455ee0d5944993e22fcec3145632b8594ba581b]] [formerly 52ba9014878f7219ab95486a132a608f5d7b8ba1 [formerly 7c89c7748e752109f2071defe92db19d360df72b] [formerly 4316bb229b7272e4fe0bc2c96f96a341103fecdd [formerly 3455ee0d5944993e22fcec3145632b8594ba581b]] [formerly e1c3f40a8b63849da0773e007583c886b973bf5e [formerly a8b0da42a0b12a8ba8e2395874f776f44f447b41] [formerly 2c013d406db00c48165da3535acd0a3000df6425 [formerly 33d5997d3e853f20341fc6ce8427dec132b20f9e]]]]]
Former-commit-id: 3ccbbc424f12e7bdae0a4b38bfe4a2115abe94f6 [formerly cc701196242af03da23d9f7aa211732a38beb748] [formerly 7eae72feb26109b6fd74c5a8106158c5875e8c31 [formerly da2d4354c3e6df779de0d13c39c546140d721470]] [formerly 3ccbbc424f12e7bdae0a4b38bfe4a2115abe94f6 [formerly cc701196242af03da23d9f7aa211732a38beb748] [formerly 7eae72feb26109b6fd74c5a8106158c5875e8c31 [formerly da2d4354c3e6df779de0d13c39c546140d721470]] [formerly e421f2ca58d8c66cc0538ddafcd7f241871fc46b [formerly 8200cf338e5f8b3698e9dc821a86aefb6eb52583] [formerly 82f3bdfe8940ae625c9f1fbb99e65f4cf1efb559 [formerly b8415a857ffab33dac6be1dfb4a0c1a8456d3970]]]] [formerly 5f2759882ae7ae21edff922b673595d7cb610bf1 [formerly 9786276ee27c56f2af663bff156be7bcf1f174bf] [formerly 8e5c998396369717935ae995f9fd01ebff00850a [formerly bca71112dc1db4b554ea1e20b06bc816985a91f3]] [formerly 5f2759882ae7ae21edff922b673595d7cb610bf1 [formerly 9786276ee27c56f2af663bff156be7bcf1f174bf] [formerly 8e5c998396369717935ae995f9fd01ebff00850a [formerly bca71112dc1db4b554ea1e20b06bc816985a91f3]] [formerly cc4e42a7f8af85971555e3a6ab7cf5624cb938d2 [formerly 85a9fa1f372ae6ce383f3ca29ab118397b9d25e7] [formerly 0b5d75ad40e8aa3b29d607852ca25d5590f092b8 [formerly 67e465bdccfa2e7bdf0285fd07640edfb4ea010d]]]]]
Former-commit-id: 5f2759882ae7ae21edff922b673595d7cb610bf1 [formerly 9786276ee27c56f2af663bff156be7bcf1f174bf] [formerly 8e5c998396369717935ae995f9fd01ebff00850a [formerly bca71112dc1db4b554ea1e20b06bc816985a91f3]]
Former-commit-id: a9a5f1aba8cc81409ac72fbb1c9c9a2424bc332a [formerly 4af1c06c0af9be0388467efbe721f7271aa7054a] [formerly baa0b10733b4951f11ece268558ce4186f16c4df [formerly 97d85a1f320e5b251cad7716906c586a2bc69dab]] [formerly 808a0932675322bdd52d0dc30a2adbfa3dc393c8 [formerly 52b7757fb5f533c13e3b257ea7ab2951953401e8] [formerly a3955cde5cc21126e886a4b855c1c802db9c551d [formerly 7900c27dc47f28476e21a5219bbbac4b5dd78c20]]]
Former-commit-id: d67c53c2ce69ade4438504bea36a3249f059df44 [formerly fba5afde8844c9a4a823cc4a18b686de95e32f3f] [formerly 7c3bd7e11998b9cc197b71ee8e1a8005c97ef3f4 [formerly a1b116b52e5a70ca7ffb67c339e13defc34dd5a9]]
Former-commit-id: 8b15c3acfb2279d3354ae994325245134b3ec889 [formerly 5689665be6776d7e54f68481bca28e525f13f334]
Former-commit-id: e1e1fa5faa37e49dca2b3f83cdbfefe7f30d2540

+++ src/mitll/langtest/server/scoring/ASRScoring.java
@@ -507,12 +507,12 @@ public class ASRScoring extends Scoring {
 
     // there are currently only two letter-to-sound classes EnglishLTS and ArabicLTS.  What about french? spanish?
     boolean isArabicScript = !(language.equalsIgnoreCase("English"));
-    if (!isArabicScript) {
+  //  if (!isArabicScript) {
       //logger.debug("using english LTS sound class since language is " + language);
-    }
+  //  }
 
     Scores scoresFromHydec = getScoresFromHydec(testAudio, sentence, configFile);
-    if (scoresFromHydec.hydecScore != -1) {
+    if (scoresFromHydec.hydecScore != -1 && false) {
       try {
         FileUtils.deleteDirectory(new File(tmpDir));
       } catch (IOException e) {
@@ -555,7 +555,7 @@ public class ASRScoring extends Scoring {
       jscoreOut = testAudio.jscore(sentence, htkDictionary, letterToSoundClass, configFile);
       float hydec_score = jscoreOut._1;
       long timeToRunHydec = System.currentTimeMillis() - then;
-      logger.info("getScoresFromHydec : got score " + hydec_score +" and took " + timeToRunHydec + " millis");
+      logger.debug("getScoresFromHydec : got score " + hydec_score +" and took " + timeToRunHydec + " millis");
 
       return new Scores(hydec_score, jscoreOut._2);
     } catch (AssertionError e) {
